refactor(DijkstraShortestPathFinderTest): Dijkstra 라이브러리를 활용한 최단경로조회 관련 기능 단위테스트 추가
